+++ kloppy/__init__.py
@@ -3,3 +3,5 @@ from .helpers import *
 from .infra import datasets
 from .domain.services.matchers.pattern import event as event_pattern_matching
 from .domain.services.state_builder import add_state
+
+__version__ = "1.5.0"